Load the NIFs in the on_load callback
This allows repetitive starting and stopping of the fast_xml
application. Since OTP 20, it is not possible to reload NIFs and doing
`erlang:load_nif` for the second time ends up in error.

Loading the NIF upon the module load instead on application startup
solves this problem.

There was one gotcha when running the tests. Cover reloads each module
with its own cover-enabled version, which has lead to calling the NIFs
`upgrade` callback, which was not implemented. Let's add this callback
and just call into the `load` callback from it.

Also the `start_test` content is not relevant anymore, so it's replaced
by testing the subsequent application starting and stopping.